[train] Validate Anyscale ScalingConfig (#2757)
https://github.com/ray-project/ray/pull/58845 added
`bundle_label_selector` to `ScalingConfig`. This PR validates that the
length of `bundle_label_selector` equals the length of `max_workers`
when users do elastic training. When we create a group of `n` workers,
we should use the first `n` bundle selectors that the user specified in
`bundle_label_selector`.

---------

Signed-off-by: Timothy Seah <tseah@anyscale.com>
Co-authored-by: Justin Yu <justinvyu@anyscale.com>