[lineage] Infer file format from file extension if ListFiles extensions are not set (#2790)
For certain read APIs like `read_csv`, `read_text`, the
`file_extensions` attribute on `ListFiles` operator is set to `None`.
This causes file format to be marked as `UNKNOWN` when data is read
using these APIs. To circumvent this, this patch adds a function `
_infer_format_from_path` to infer the file format from the real file
extension. If it cannot be inferred from the real file extension as
well, then the format is marked as `UNKNOWN`.