[data][Autoscaler][1/N] Add Util threshold to prevent scaling (#2776)
The rate-based autoscaler assumes that the throughput of the bottleneck operator improves if it has more resources and always scales nodes for the bottleneck operator. However, this assumption doesn't always hold (e.g., if the user has manually specified concurrency).

To fix this issue, this PR only scales up resources if the utilization is high for any resource type.

---------

Signed-off-by: iamjustinhsu <jhsu@anyscale.com>