[Data] Enforce `SupportsClusterAutoscaling` as a subset of `PhysicalOperator` (#2835)
Previously, the rate-based autoscaler only scaled operators that
implement the `SupportsClusterAutoscaling` interface, which was intended
to be a subset of `PhysicalOperator`. However,
https://github.com/ray-project/ray/pull/59331 removed a method from
`PhysicalOperator`, causing `SupportsClusterAutoscaling` to no longer be
a strict subset of the `PhysicalOperator` interface.

To prevent this class of issue in the future, this PR adds an assertion
to explicitly enforce the interface relationship.

---------

Signed-off-by: Balaji Veeramani <bveeramani@berkeley.edu>