[train] Renaming DatasetsSetupCallback to DatasetsCallback (#2813)
This is the analogous Rayturbo PR for the [OSS
PR](https://github.com/ray-project/ray/pull/59423) that renames the OSS
DatasetsSetupCallback to DatasetsCallback. This PR renames the Rayturbo
DatasetsSetupCallback to DatasetsCallback to align with OSS changes.

Note: The OSS renaming is included in this PR (duplicated from above
linked OSS PR) to allow for tests to pass but the OSS PR will be merged
first.

Signed-off-by: JasonLi1909 <jasli1909@gmail.com>