[Data] Remove Anyscale-only `min_max_resource_usage` implementation (#2863)
To fix a GPU allocation bug,
https://github.com/ray-project/ray/pull/59632 added
`min_max_resource_usage` implementations to OSS Ray. This has caused
some of the RayTurbo tests to fail because the implementations are
slightly different.

To fix the failing tests, and also reduce the maintenance burden of
duplicate implementation, this PR removes the RayTurbo implementations.